refactor: rename room.py to abstract_room_card.py to match internal class
