+++ backend/src/main/game/room/abstract_room_card.py
