+++ backend/src/test/game/room/test_constructed_room.py
@@ -1,10 +1,9 @@
 import pytest
-
 from backend.src.main.game.monster.concrete_monster_cards.cutthroat import Cutthroat
 from backend.src.main.game.monster.values import NumberedRoomTileValues, DungeonCardValues
+from backend.src.main.game.room.abstract_room_card import Tile
 from backend.src.main.game.room.concrete_room_cards.den import Den
 from backend.src.main.game.room.constructed_room import ConstructedRoom
-from backend.src.main.game.room.room import Tile
 
 
 def test_numbered_tile_is_numbered_tile():
